cleanup my_dataset code
