+++ my_dataset.py
@@ -1,19 +1,14 @@
-import torch
-from torch.utils.data import Dataset, DataLoader, random_split
-from torch.utils.data.sampler import SubsetRandomSampler
-import enum
 import os
 import copy
 import cv2
+import enum
 import numpy as np
 import torchvision
-import matplotlib.pyplot as plt
 from PIL import Image
+from torch.utils.data import Dataset, DataLoader, random_split
 import util
 
 train_path = '../bit-bots-ball-dataset-2018/train'
-valid_path = '../bit-bots-ball-dataset-2018/valid'
-negative_path = '../bit-bots-ball-dataset-2018/negative'
 test_path = '../bit-bots-ball-dataset-2018/test'
 
 
@@ -53,8 +48,8 @@ def initialize_loader(batch_size, num_workers=64, shuffle=True):
                              num_workers=num_workers,
                              shuffle=shuffle)
 
-    print('train dataset: # images {:>6}, # robots {:>6}, # balls {:>6}'.format(
-        len(train_dataset),
+    print('full dataset: # images {:>6}, # robots {:>6}, # balls {:>6}'.format(
+        len(full_dataset),
         full_dataset.num_robot_labels,
         full_dataset.num_ball_labels
     ))
@@ -74,57 +69,9 @@ def initialize_loader(batch_size, num_workers=64, shuffle=True):
     return (train_loader, valid_loader, test_loader), (full_dataset, test_dataset)
 
 
-def draw_bounding_boxes(img, bbxs, colour):
-    '''
-    :param img: rgb torch image
-    :param bbxs:
-    :param colour:
-    :return:
-    '''
-    img = util.torch_to_cv(img)
-    img = img.copy()  # cv2 seems to like copies to draw rectangles on
-
-    for bbx in bbxs:
-        if isinstance(bbx, str):
-            return util.cv_to_torch(img)
-        pt0 = (int(bbx[0]), int(bbx[1]))
-        pt1 = (int(bbx[2]), int(bbx[3]))
-        img = cv2.rectangle(img, pt0, pt1, colour, 1)
-    return util.cv_to_torch(img)
-
-
-def display_image(to_plot):
-    '''
-    :param to_plot: list of tuples of the form (img [(cxhxw) numpy array], cmap [str], title [str])
-    '''
-    fig, ax = plt.subplots(3, 2, figsize=(8, 10))
-    for i, plot_info in enumerate(to_plot):
-        img = util.torch_to_cv(plot_info[0])
-        cmap = plot_info[1]
-        title = plot_info[2]
-
-        ax[i // 2, i % 2].imshow(img, cmap=cmap)
-        ax[i // 2, i % 2].set_title(title)
-    plt.show()
-
-
-def stream_image(img, wait, scale):
-    img = util.torch_to_cv(img)
-    width, height, _ = img.shape
-    img = cv2.resize(img, (height * scale, width * scale))
-    img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-    cv2.imshow('my_window', img)
-    cv2.waitKey(wait)
-
-
-def read_image(path):
-    # using opencv imread crashes Pytorch DataLoader for some reason
-    return Image.open(path)
-
-
 class Label(enum.Enum):
     '''
-    the values correspond to which output neuron should be activated
+    Defines output layers of model
     '''
     BALL = 0
     ROBOT = 1
@@ -196,7 +143,7 @@ class MyDataSet(Dataset):
         '''
         img_path = self.img_paths[index]
         bounding_boxes = self.bounding_boxes[img_path]
-        img = read_image(img_path)
+        img = util.read_image(img_path)
 
         height, width, _ = np.array(img).shape
         # the final mask will have no channels but we need 3 to convert to PIL image to apply transformation
@@ -227,7 +174,7 @@ class MyDataSet(Dataset):
 
     def get_bounding_boxes(self, index):
         img_path = self.img_paths[index]
-        img = read_image(img_path)
+        img = util.read_image(img_path)
 
         # we need to scale bounding boxes since we applied a transformation
         height, width, _ = np.shape(img)
@@ -246,5 +193,5 @@ class MyDataSet(Dataset):
         for ind in range(len(self)):
             img, _, _ = self[ind]
             bbxs = self.get_bounding_boxes(ind)
-            img = draw_bounding_boxes(img, bbxs, 255)
-            stream_image(img, delay, scale)
+            img = util.draw_bounding_boxes(img, bbxs, 255)
+            util.stream_image(img, delay, scale)
